engine/requirement.txt
openai
tiktoken
pydantic
python-dotenv
selenium
requests
requests-toolbelt
beautifulsoup4
lxml